Import appropriate portions of the UniStaker codebase
The initial commit starts with a stripped down and slightly updated version of the UniStaker
codebase. In particular, the V3FactoryOwner and all related tests and scripts was removed. It
is specific only to the Uniswap ecosystem and irrelevant to gov staker. Similiarly, the integration
tests, which are Uniswap specifc, were removed, as were the Uniswap related submodules. The forge
standard library was updated, as was OpenZeppelin, and minor changes were made to accomodate slight
differences in these libraries, such as marking some tests as view and reducing the number of iterations
in certain tests to avoid running out of gas. The Solidity version was moved up to 0.8.26 and
constraints on the compiler pragmas were loosened where needed.