Add a configurable claim fee to governance staker
The claim fee is intentionally set as a fixed amount. The primary purpose of this
fee is to act as a deterrent to "deposit sybiling", whereby a staker might spread
funds out across many deposits in an attempt to reduce the liklihood or frequency
of being "bumped" if his earning power decreases. Allowing the DAO to set a small,
fixed fee would decrease the effective yield of small deposits without materially
impacting a normal user.