Implement and test a `transferFrom` based reward notifier
Create and test a reward notifier contract that uses `transferFrom` to pull reward tokens
from a pre-approved source address on a regular cadence. The source, reward amount, and
reward interval are configurable by the contract's owner. It is assumed the source has
used `approve` to allow the notifier contract to pull the tokens. The notify method is open
and can be called by anyone. It does not provide a tip to the caller: the assumption is that
enough stakers will be incentivized such that someone will call it. DAOs can also set up a
bot do this automatically.